Changed fact and the corresponding files to use a Token instead of a string.
Added a contructor to Token that takes only a string and sets the type to
uknown. This was needed by MovieCharacterFactExtractor, as the Tokenizer
is never run of the input strings, meaning the type of them is unkown.

The constructor may need to be changed later? Not entirely sure if
MovieCharacterFact extractor will remain relevant with the current
architecture, meaning there would be no need to change it.