Altered the main & other files
Changed the main to work with the current querey code and the current match fact code. Currently it reflects a single structure, and looks for the corresponding Fact in a 'demo' pool. Once we have the real pool, we can simply expand the table of known structures to enable us to search for more Facts.